feat(testworkflows): move TestWorkflow Init Process outside TCL
