+++ cmd/testworkflow-init/data/types.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/kubeshop/testkube/cmd/tcl/testworkflow-init/constants"
+	"github.com/kubeshop/testkube/cmd/testworkflow-init/constants"
 )
 
 type TestWorkflowStatus string
