+++ pkg/tcl/testworkflowstcl/testworkflowcontroller/notification.go
@@ -12,7 +12,7 @@ import (
 	"encoding/json"
 	"time"
 
-	"github.com/kubeshop/testkube/cmd/tcl/testworkflow-init/data"
+	"github.com/kubeshop/testkube/cmd/testworkflow-init/data"
 	"github.com/kubeshop/testkube/pkg/api/v1/testkube"
 	"github.com/kubeshop/testkube/pkg/log"
 )
